remove java version from pom change signature of stocksystem method

+++ org.dieschnittstelle.jee.esa.ejb.ejbmodule.erp/src/main/java/org/dieschnittstelle/jee/esa/ejb/ejbmodule/erp/StockSystemLocal.java
@@ -65,7 +65,7 @@ public interface StockSystemLocal {
 	 * @param product
 	 * @return
 	 */
-	public List<Integer> getPointsOfSale(IndividualisedProductItem product);
+	public List<Long> getPointsOfSale(IndividualisedProductItem product);
 
 
 	/**
