+++ org.dieschnittstelle.jee.esa.ejb.ejbmodule.erp/src/main/java/org/dieschnittstelle/jee/esa/ejb/ejbmodule/erp/StockSystemRemote.java
@@ -62,6 +62,6 @@ public interface StockSystemRemote {
 		 * @param product
 		 * @return
 		 */
-		public List<Integer> getPointsOfSale(IndividualisedProductItem product);
+		public List<Long> getPointsOfSale(IndividualisedProductItem product);
 
 }
